+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
-- Added a result attribute to the handler object (HandlerCoroutine)
+- Added a result attribute to the handler (HandlerCoroutine) and Job object
+- When an exception is encountered, add the exception to the exception attribute of the corresponding handler or Job
 
 ## [0.1.3] - 2023-01-05